Rename file appropriately.
